[FEAT] adding search component to store page
