+++ src/components/pages/Store/index.tsx
@@ -1,30 +1,18 @@
 import React, { ChangeEvent, useEffect, useState } from "react";
-import { Store, useGetStore } from "../../../hooks/stores/useGetStore";
+
 import { useSearchStore } from "../../../store";
+import ContentWrapper from "../../Content/ContentWrapper";
 import Map from "../../Map";
+import Search from "../../Search";
 
 const StorePage = () => {
-  const [text, setText] = useState("");
-  const { data } = useGetStore();
   const {
     searchResult,
     searchText,
     initResult,
     setSearchResult,
-    setSearchText,
     setInitResult,
   } = useSearchStore();
-
-  const inputHandler = (event: ChangeEvent<HTMLInputElement>) => {
-    setText(event.target.value);
-    setSearchText(event.target.value);
-  };
-
-  const searchHandler = () => {
-    setSearchResult(
-      data?.filter((item: Store) => item.CMPNM_NM?.includes(text))
-    );
-  };
 
   useEffect(() => {
     setInitResult();
@@ -36,9 +24,10 @@ const StorePage = () => {
 
   return (
     <>
-      <button onClick={searchHandler}>검색</button>
-      <input type="text" onChange={inputHandler} />
-      <Map initStore={searchText === "" ? initResult : searchResult} />
+      <ContentWrapper>
+        <Search />
+        <Map initStore={searchText === "" ? initResult : searchResult} />
+      </ContentWrapper>
     </>
   );
 };
